requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.24.0,<0.28.0
python-dotenv>=1.0.0
supabase>=2.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
loguru>=0.7.0
python-telegram-bot>=20.0,<21.0 